requirements.txt
telethon==1.24.0
jinja2>=2.11.3
PyYAML>=5.4.1
cryptg==0.2.post2
Pillow>=8.3.2
feedgen>=0.9.0
python-magic>=0.4.24
python-socks[asyncio]
python-magic-bin